extract conversation manager out of /agent dir
Summary:

Test Plan:
