+++ modular-agent-fullstack/backend/src/agent/my_local_agent/route.py
@@ -13,13 +13,13 @@ from openinference.semconv.trace import SpanAttributes
 from opentelemetry import trace
 from rich import print
 
-from .conversation import ConversationManager
 from .examples import get_weather, get_weather_conditions
 import traceback
 
 from src.database.db import DatabaseManager, DatabaseUtils
 from src.models import Conversation, Role
 from src.logging_config import LOGGING_CONFIG
+from src.conversation import ConversationManager
 
 
 # Configure logging
